Added function to calculate chromatic symmetric function of a graph.
Can calculate the CSF in several different symmetric function bases, including the power sum, monomial, elementary, homogeneous, and Schur bases. It's not terribly efficient, but it gets the job done. Note that this uses Theorem 2.5 from Richard Stanley's original paper "A Symmetric Function Generalization of the Chromatic Polynomial of a Graph."