Added subset generator.
Particulars:
Generates fixed-size subsets of edge indices with strong minimal change property. Such subsets have the property that the difference between one subset and the next is the replacement of one element with another. This is a Hamiltonian circuit of the subsets. The generator returns an array of 'deltas' which are structs that record which element needs to be replaced, and with what, from subset to subset. The elements are indices to an edge array: between each subset of edge indices, we need to remove one edge and add one edge, so this makes it easy to iterate over the subsets of edges of fixed size (a crucial part of the algorithm we're trying to implement).