Bug fix to make k-slice (i.e. partial CSF), as well as tree-removal correct.
The CSF calculation itself was fine, but the edge subsets it was being passed by the subset generator caused it to add vertices to the trees, which of course messed up the CSF calculation. In addition, there were some bugs in the code which removed trees with different k-slices, so I changed to keep trees with the same k-slice rather than remove trees with different k-slices.