Reduced memory-intensiveness of the program at a cost of about 10% more runtime.
Implemented a function that builds/destroys the adjacency matrix of the tree on the fly from a lelvel sequence. Since the level sequence is an array of n integers, we can save a lot of memory by storing the level sequence rather than the entire adjacency matrix. Now, whenever we need the adjacency matrix, we simply build it, then destroy it when it is no longer needed.